requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
chromedriver-autoinstaller==0.4.0
click==8.1.3
cssselect==1.2.0
exceptiongroup==1.1.1
fastapi==0.95.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.10.0
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
huggingface-hub==0.13.2
idna==3.4
importlib-metadata==6.0.0
jieba3k==0.35.1
joblib==1.2.0
livereload==2.6.3
lxml==4.9.2
multidict==6.0.4
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.21.6
openai==0.27.2
outcome==1.2.0
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
pydantic==1.10.7
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
ratelimiter==1.2.0.post0
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
selenium==4.8.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
starlette==0.26.1
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.13.2
torch==1.13.1
tornado==4.5.3
tqdm==4.65.0
transformers==4.27.1
trio==0.22.0
trio-websocket==0.10.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchdog==3.0.0
webdriver-manager==3.8.5
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0
